requirements.txt
psycopg2
statistics
mmh3
python-dotenv